rename Internal to lagacy
